Joi hates undefined strings
Problem: according to its documentation all arguments in Joi are
considered optional. This is appearently incorrect, at least for
strings. It is currently unknown how to get joi to accept undefined
values as acceptable for strings which were meant to be optional
arguments.

to try this:

run `yarn test ./test/paratii.eth.vids.js` to try to get joi to handle
undefined strings for `id`, `ipfsHashOrig`, etc.

These seemed helpful, but did not fix the problem:
- [https://github.com/hapijs/joi/issues/482](https://github.com/hapijs/joi/issues/482)
- [https://github.com/hapijs/joi/issues/251](https://github.com/hapijs/joi/issues/251)